document-classification-agent/requirements.txt
flask
pytesseract
pdf2image
Pillow
python-dotenv
openai
